+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.CypressTests/cypress/e2e/smoke/smoke-create-presumption-project-to-check-answers-page.cy.ts
@@ -5,13 +5,13 @@ import createProjectPage from "cypress/pages/createProject/createProjectPage";
 import homePage from "cypress/pages/homePage";
 import summaryPage from "cypress/pages/task-summary-base";
 
-    describe("Smoke Test Create an individual project - (Up Until Check your answers Page)", () => {
+    describe("Smoke Test Create an individual Presumption project - (Up Until Check your answers Page)", () => {
         beforeEach(() => {
             cy.login({ role: ProjectRecordCreator });
             cy.visit('/');
         });
 
-        it("Should Smoke Test go through create a project wizard happy path up until Check your answers page", { tags: ['smoke'] }, () => {
+        it("Should create a presumption project, up until Check your answers page", { tags: ['smoke'] }, () => {
             const temporaryProjectId = dataGenerator.generateTemporaryId(25);
             const schoolName = dataGenerator.generateSchoolName();
             const TestTrn = "TR00111";
